Instructions for Day 1
As we go through each part of this document, there is accompanying code for the part. Not always but mostly. Most of the code assumes you have a Node.JS environment setup. It is possible to perform some of the advanced code using a browser. The instructor will provide a template file for that.